+++ src/Game.ts
@@ -536,6 +536,9 @@ export class Game {
 
 	/**
 	 * ゲームにjoinしているプレイヤーIDの一覧。
+	 *
+	 * 歴史的経緯により `undefined` は含まれないことに注意。
+	 * `undefined` は特殊はプレイヤーIDであり通常 join しないため、ゲーム開発者がこの仕様を考慮する必要はない。
 	 */
 	joinedPlayerIds: string[];
 
@@ -842,7 +845,7 @@ export class Game {
 		this._mainParameter = undefined;
 		this._configuration = gameConfiguration;
 		this._assetManager = new AssetManager(this, gameConfiguration.assets, gameConfiguration.audio, gameConfiguration.moduleMainScripts);
-		this._moduleManager = new ModuleManager(this._runtimeValueBase, this._assetManager);
+		this._moduleManager = undefined!;
 
 		const operationPluginsField = gameConfiguration.operationPlugins || [];
 		this.operationPluginManager = new OperationPluginManager(this, param.operationPluginViewInfo || null, operationPluginsField);
@@ -1351,6 +1354,7 @@ export class Game {
 
 		this._isTerminated = false;
 		this.vars = {};
+		this._moduleManager = new ModuleManager(this._runtimeValueBase, this._assetManager);
 
 		this.surfaceAtlasSet.destroy();
 		this.surfaceAtlasSet = new SurfaceAtlasSet({ resourceFactory: this.resourceFactory });
@@ -1475,6 +1479,7 @@ export class Game {
 		this._postTickTasks = [];
 		this.surfaceAtlasSet.destroy();
 		this.surfaceAtlasSet = undefined!;
+		this._moduleManager = undefined!;
 	}
 
 	/**
@@ -1624,7 +1629,9 @@ export class Game {
 	 * @ignore
 	 */
 	_handleJoinEvent(event: JoinEvent): void {
-		if (this.joinedPlayerIds.indexOf(event.player.id) !== -1) return;
+		if (!event.player.id || this.joinedPlayerIds.indexOf(event.player.id) !== -1) {
+			return;
+		}
 		this.joinedPlayerIds.push(event.player.id);
 	}